api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-jose[cryptography]"
]